Check for non existent file in copy_shared_libraries.sh

+++ src/copy_shared_libraries.sh
@@ -53,6 +53,10 @@ copy_if_newer() {
   local libname
   libname=$(basename "${source}")
 
+  if [[ ! -e "${source}" ]] ; then
+    return
+  fi
+
   if [[ ! -s "${destdir}/${libname}" || "${source}" -nt "${destdir}/${libname}" ]] ; then
     echo "Copying ${source}"
     cp -f "${source}" "${destdir}"
